+++ src/compas_viewer/components/sceneform.py
@@ -66,7 +66,7 @@ class Sceneform(QTreeWidget):
         return self.viewer.scene
 
     def update(self):
-        self.clear()
+        self.clear()  # TODO: do not clear when objects are same.
         self.checkbox_columns = {}
 
         for node in self.scene.traverse("breadthfirst"):
@@ -94,6 +94,16 @@ class Sceneform(QTreeWidget):
             widget = QTreeWidgetItem(parent_widget, strings)
             widget.node = node
             widget.setSelected(node.is_selected)
+            if node.is_selected:
+
+                def expand(node):
+                    if node.attributes.get("widget"):
+                        node.attributes["widget"].setExpanded(True)
+                        if node.parent and not node.parent.is_root:
+                            expand(node.parent)
+
+                expand(node.parent)
+
             widget.setFlags(widget.flags() | Qt.ItemIsUserCheckable | Qt.ItemIsSelectable | Qt.ItemIsEnabled)
 
             for col, col_data in self.checkbox_columns.items():
@@ -113,7 +123,7 @@ class Sceneform(QTreeWidget):
             for node in self.scene.objects:
                 node.is_selected = node in selected_nodes
                 if self.callback and node.is_selected:
-                    self.callback(node)
+                    self.callback(self, node)
 
             self.viewer.ui.sidebar.update()
 
@@ -122,7 +132,7 @@ class Sceneform(QTreeWidget):
     def on_item_selection_changed(self):
         for item in self.selectedItems():
             if self.callback:
-                self.callback(item.node)
+                self.callback(self, item.node)
 
     def adjust_column_widths(self):
         for i in range(self.columnCount()):