+++ CHANGELOG.md
@@ -9,6 +9,54 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Changed
+
+### Removed
+
+
+## [1.2.3] 2024-07-03
+
+### Added
+
+* Added `scripts/treeform.py` example.
+
+### Changed
+
+### Removed
+
+
+## [1.2.2] 2024-07-02
+
+### Added
+
+### Changed
+
+* Fixed issue of `SceneForm` collapsing when new object is selected.
+* Toggling `show` of Group or objects will hide/show all its descendents.
+* Moved assets related files to `assets` folder.
+* Fixed a bug to allow changing `Viewer.renderer.view` before `Viewer.show()`.
+* Fixed a bug of grid not showing while the scene is empty.
+
+### Removed
+
+* Removed `is_locked` from `compas_viewer.scene.ViewerSceneObject`.
+
+
+## [1.2.1] 2024-06-26
+
+### Added
+
+### Changed
+
+* Fixed missing shader files.
+
+### Removed
+
+
+## [1.2.0] 2024-06-26
+
+### Added
+
 * Added `compas_viewer.components.dialog` component which handle camera setting popup window.
 * Added `compas_viewer.components.combo_box` component.
 * Added `ViewComboBox` component handle viewmode change.